+++ SCALA_VERSION
@@ -1,0 +1,1 @@
+3.3.1
