+++ wvlet-ui/index.html
@@ -23,7 +23,6 @@
 </head>
 <body>
 <div id="app"></div>
-<div id="editor" style="width: 800px; height: 250px; border: 1px solid grey"></div>
 <script type="module" src="/main.js"></script>
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4" crossorigin="anonymous"></script>
 </body>
