+++ wvlet-ui/package.json
@@ -15,6 +15,6 @@
   },
   "devDependencies": {
     "@scala-js/vite-plugin-scalajs": "1.0.0",
-    "vite": "^4.3.9"
+    "vite": "^4.5.0"
   }
 }
